requirements.txt
fastapi
uvicorn
opencv-python
pyzbar
python-multipart
jinja2
sqlalchemy
numpy
requests
fpdf2
openpyxl
